feat(session): consolidate date buttons and implement mobile layout
